reafact: move functions
